requirements.txt
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pytesseract>=0.3.10
easyocr>=1.7.0
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0
autocorrect>=2.6.1
textdistance>=4.6.0
regex>=2023.10.3
